% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets-catchments.R
\name{rf.helper.copy_file}
\alias{rf.helper.copy_file}
\title{Copies NHD files with \code{vpu} to \code{new_dir}.}
\usage{
rf.helper.copy_file(vpu, files, new_dir)
}
\arguments{
\item{vpu}{VPU identifier}

\item{files}{List of files to scan}

\item{new_dir}{Directory to copy matching files into}
}
\value{
Path(s) of copied file(s)
}
\description{
Copies NHD files with \code{vpu} to \code{new_dir}.
}
\keyword{internal}
