% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets-burnlines.R
\name{rf.targets.burnline_events}
\alias{rf.targets.burnline_events}
\title{Burnline Events Processing Target}
\usage{
rf.targets.burnline_events(nhd_gdb, rf_vaa, dir_data)
}
\arguments{
\item{nhd_gdb}{Path to NHDPlus Seamless GeoDatabase}

\item{rf_vaa}{NHD Value-added attributes table}

\item{dir_data}{Data output directory}
}
\value{
Path to outputted burn line events file for CONUS
}
\description{
Processes burn line events according to step 4.3.1
}
