% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets-catchments.R
\name{rf.targets.clean_catchment}
\alias{rf.targets.clean_catchment}
\title{Clean Catchments Target}
\usage{
rf.targets.clean_catchment(cat_info, simplify_keep = 0.2)
}
\arguments{
\item{cat_info}{\code{data.frame} containing path information for catchments and flowlines}

\item{simplify_keep}{Tolerance value for weighted Visvalingam simplification}
}
\value{
Path to cleaned catchment output (\code{cat_info$outfile}).
}
\description{
Processes catchments according to steps 4.2.1 - 4.2.3.
}
