% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets-flowlines.R
\name{rf.targets.clean_flowlines}
\alias{rf.targets.clean_flowlines}
\title{Clean Flowlines Target}
\usage{
rf.targets.clean_flowlines(
  flowlines_path,
  rf_cat_paths,
  rf_ble_path,
  rf_vaa,
  rf_enhd,
  dir_cleaned
)
}
\arguments{
\item{flowlines_path}{Path to NHD flowlines file}

\item{rf_cat_paths}{Path to \emph{reference} catchment outputs}

\item{rf_ble_path}{Path to CONUS burn line events file}

\item{rf_vaa}{NHDPlus-VAA table}

\item{rf_enhd}{E2NHDPlus River Attributes table}

\item{dir_cleaned}{Cleaned output directory}
}
\value{
Path to cleaned flowlines
}
\description{
Processes flowlines according to step 4.3.2
}
