% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets-waterbodies.R
\name{rf.targets.clean_waterbodies}
\alias{rf.targets.clean_waterbodies}
\title{Clean Waterbodies Target}
\usage{
rf.targets.clean_waterbodies(waterbodies_path, dir_cleaned)
}
\arguments{
\item{waterbodies_path}{Path to unmodified NHDWaterbody file}

\item{dir_cleaned}{Directory to output cleaned waterbody geometry}
}
\value{
Path to cleaned waterbodies
}
\description{
Processes waterbodies according to steps 4.1.1 - 4.1.6.
}
