% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets-catchments.R
\name{rf.targets.rectify_catchment_borders}
\alias{rf.targets.rectify_catchment_borders}
\title{Catchment VPU Rectification Target}
\usage{
rf.targets.rectify_catchment_borders(
  cat_cleaned_paths,
  vpu_topology,
  dir_reference
)
}
\arguments{
\item{cat_cleaned_paths}{Paths to cleaned catchment files}

\item{vpu_topology}{\code{data.frame} containing topology definition for VPUs}

\item{dir_reference}{Output directory for reference features}
}
\value{
List of catchment output paths
}
\description{
Processes VPU rectification of \emph{cleaned} catchments according to step 4.2.4.
}
\note{
This process is highly data-dependent, so we must process it sequentially.
}
