% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets-catchments.R
\name{rf.targets.reference_catchments}
\alias{rf.targets.reference_catchments}
\title{Catchments Reference Output Target}
\usage{
rf.targets.reference_catchments(rf_cat_path)
}
\arguments{
\item{rf_cat_path}{Path to VPU-rectified catchment}
}
\value{
Path to outputted reference catchment
}
\description{
Processes remaining catchment steps according to step 4.2.5.
}
\section{FIXME}{
 Snap to underlying grid with size of .0009?
}

