% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets-flowlines.R
\name{rf.targets.reference_flowlines}
\alias{rf.targets.reference_flowlines}
\title{Reference Flowlines Target}
\usage{
rf.targets.reference_flowlines(nhd, vpu, rf_enhd_comid, dir_reference)
}
\arguments{
\item{nhd}{Path to \emph{cleaned} flowlines}

\item{vpu}{Current VPU for \code{nhd}.}

\item{rf_enhd_comid}{COMIDs with E2NHDPlus river attributes}

\item{dir_reference}{Reference output directory}
}
\value{
Path to reference flowlines
}
\description{
Processes flowlines according to step 4.3.3
}
