% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets-waterbodies.R
\name{rf.targets.reference_waterbodies}
\alias{rf.targets.reference_waterbodies}
\title{Process Reference Waterbodies Target}
\usage{
rf.targets.reference_waterbodies(wb_clean, wb_vpu, wb_reference_dir)
}
\arguments{
\item{wb_clean}{Path to cleaned waterbodies file}

\item{wb_vpu}{The corresponding VPU of \code{wb_clean}}

\item{wb_reference_dir}{Directory to output reference waterbodies}
}
\value{
Path to reference waterbodies
}
\description{
Processes \emph{cleaned} waterbodies according to steps 4.1.7 - 4.1.9.
}
\section{FIXME}{
 On/off network waterbodies based on NHD flowlines?
}

